feat: improved (i.e. tractable) initilization of gene sets to test
The original method created all combinations of all mutated genes.
This quickly prooved to be too many combinations and took too
long and too much RAM

The new method accounts for the mutation frequency of each gene,
the test (mut. ex. or co-mut.), and the seed genes. It is now
much faster, more efficient, and can actually be used with a
real data set.

(Tests are still needed)