fix: swap_an_edge2() no longer randomly creates a double edge
The problem was in the samples of available edges for the second
edge. If only one edge was available, then this was provided
as an integer to the first paramter of sample(). When sample()
is given an integer, it instead uses seq_along(1, n) to sample
from. Therefore, I have added a check to how many avilable
edges there are - if there is only one, I just use that edge.